docs.requirements.txt
mkdocs-material>=9.0.0
mkdocs-macros-plugin>=0.7.0
mkdocs-drawio-file==1.3.0
mkdocs-table-reader-plugin==1.2
mkdocs-minify-plugin==0.6.2
mkdocs-redirects==1.2.0
mkdocs-glightbox>=0.3.2
pymdown-extensions>=9.9.2

surquest-utils-config-formatter>=0.0.1rc4
surquest-GCP-bq-grid>=0.0.1rc4